refactor: 사용자 출석률 조회 DTO int Casting 제거

+++ src/main/java/com/kgu/studywithme/member/infrastructure/query/MemberInformationRepository.java
@@ -169,7 +169,7 @@ public class MemberInformationRepository implements MemberInformationRepositoryA
                 .select(
                         new QAttendanceRatio(
                                 studyAttendance.status,
-                                studyAttendance.count().intValue()
+                                studyAttendance.count()
                         )
                 )
                 .from(studyAttendance)
